requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
joblib==1.0.1
numpy==1.20.1
psycopg2==2.8.6
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.1
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
whitenoise==5.2.0
